requirements.txt
pikepdf==2.12.2
Pillow==8.2.0
